SimulationData: added left-/rightSteerings, changed memory to kb from bytes, switched to uint62_t for most fields

+++ src/common/data/simulationframedata.cpp
@@ -37,17 +37,17 @@ namespace Common {
     {
     }
 
-    void FrameMemory::setMemoryBeforeFrame(uint32_t memBeforeFrame)
+    void FrameMemory::setMemoryBeforeFrame(uint64_t memBeforeFrame)
     {
         before = memBeforeFrame;
     }
 
-    void FrameMemory::setMemoryDuringFrame(uint32_t memDuringFrame)
+    void FrameMemory::setMemoryDuringFrame(uint64_t memDuringFrame)
     {
         during = memDuringFrame;
     }
 
-    void FrameMemory::setMemoryAfterFrame(uint32_t memAfterFrame)
+    void FrameMemory::setMemoryAfterFrame(uint64_t memAfterFrame)
     {
         after = memAfterFrame;
     }
@@ -57,27 +57,27 @@ namespace Common {
         return "{" + std::to_string(frame) + "," + std::to_string(before) + "," + std::to_string(during) + "," + std::to_string(after) + "}";
     }
 
-    uint32_t FrameMemory::sum() const
+    uint64_t FrameMemory::sum() const
     {
         return before + during + after;
     }
 
-    uint32_t FrameMemory::average() const
+    uint64_t FrameMemory::average() const
     {
         return sum() / parts;
     }
 
-    uint32_t FrameMemory::minimum() const
+    uint64_t FrameMemory::minimum() const
     {
-        uint32_t min = before;
+        uint64_t min = before;
         min = std::min(min, during);
         min = std::min(min, after);
         return min;
     }
 
-    uint32_t FrameMemory::maximum() const
+    uint64_t FrameMemory::maximum() const
     {
-        uint32_t max = before;
+        uint64_t max = before;
         max = std::max(max, during);
         max = std::max(max, after);
         return max;
